requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.9.1     # via bs4
bs4==0.0.1                # via notion
cached-property==1.5.1    # via notion
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
commonmark==0.9.1         # via notion
dictdiffer==0.8.1         # via notion
idna==2.10                # via requests
notion==0.0.25            # via -r requirements.in
python-dotenv==0.14.0     # via -r requirements.in
python-slugify==4.0.1     # via notion
pytz==2020.1              # via tzlocal
requests==2.24.0          # via notion
schedule==0.6.0           # via -r requirements.in
soupsieve==2.0.1          # via beautifulsoup4
text-unidecode==1.3       # via python-slugify
tzlocal==2.1              # via notion
urllib3==1.25.9           # via requests
